chore: simplify composable using a server plugin
